Remove DOM injection for attributeName
