+++ src/renderers/dom/shared/SVGDOMPropertyConfig.js
@@ -11,147 +11,6 @@
 
 'use strict';
 
-var DOMProperty = require('DOMProperty');
-
-var {HAS_STRING_BOOLEAN_VALUE} = DOMProperty.injection;
-
-var NS = {
-  xlink: 'http://www.w3.org/1999/xlink',
-  xml: 'http://www.w3.org/XML/1998/namespace',
-};
-
-/**
- * This is a list of all SVG attributes that need special casing,
- * namespacing, or boolean value assignment.
- *
- * When adding attributes to this list, be sure to also add them to
- * the `possibleStandardNames` module to ensure casing and incorrect
- * name warnings.
- *
- * SVG Attributes List:
- * https://www.w3.org/TR/SVG/attindex.html
- * SMIL Spec:
- * https://www.w3.org/TR/smil
- */
-var ATTRS = [
-  'accent-height',
-  'alignment-baseline',
-  'arabic-form',
-  'baseline-shift',
-  'cap-height',
-  'clip-path',
-  'clip-rule',
-  'color-interpolation',
-  'color-interpolation-filters',
-  'color-profile',
-  'color-rendering',
-  'dominant-baseline',
-  'enable-background',
-  'fill-opacity',
-  'fill-rule',
-  'flood-color',
-  'flood-opacity',
-  'font-family',
-  'font-size',
-  'font-size-adjust',
-  'font-stretch',
-  'font-style',
-  'font-variant',
-  'font-weight',
-  'glyph-name',
-  'glyph-orientation-horizontal',
-  'glyph-orientation-vertical',
-  'horiz-adv-x',
-  'horiz-origin-x',
-  'image-rendering',
-  'letter-spacing',
-  'lighting-color',
-  'marker-end',
-  'marker-mid',
-  'marker-start',
-  'overline-position',
-  'overline-thickness',
-  'paint-order',
-  'panose-1',
-  'pointer-events',
-  'rendering-intent',
-  'shape-rendering',
-  'stop-color',
-  'stop-opacity',
-  'strikethrough-position',
-  'strikethrough-thickness',
-  'stroke-dasharray',
-  'stroke-dashoffset',
-  'stroke-linecap',
-  'stroke-linejoin',
-  'stroke-miterlimit',
-  'stroke-opacity',
-  'stroke-width',
-  'text-anchor',
-  'text-decoration',
-  'text-rendering',
-  'underline-position',
-  'underline-thickness',
-  'unicode-bidi',
-  'unicode-range',
-  'units-per-em',
-  'v-alphabetic',
-  'v-hanging',
-  'v-ideographic',
-  'v-mathematical',
-  'vector-effect',
-  'vert-adv-y',
-  'vert-origin-x',
-  'vert-origin-y',
-  'word-spacing',
-  'writing-mode',
-  'x-height',
-  'xlink:actuate',
-  'xlink:arcrole',
-  'xlink:href',
-  'xlink:role',
-  'xlink:show',
-  'xlink:title',
-  'xlink:type',
-  'xml:base',
-  'xmlns:xlink',
-  'xml:lang',
-  'xml:space',
-];
-
-var SVGDOMPropertyConfig = {
-  Properties: {
-    autoReverse: HAS_STRING_BOOLEAN_VALUE,
-    externalResourcesRequired: HAS_STRING_BOOLEAN_VALUE,
-    preserveAlpha: HAS_STRING_BOOLEAN_VALUE,
-  },
-  DOMAttributeNames: {
-    autoReverse: 'autoReverse',
-    externalResourcesRequired: 'externalResourcesRequired',
-    preserveAlpha: 'preserveAlpha',
-  },
-  DOMAttributeNamespaces: {
-    xlinkActuate: NS.xlink,
-    xlinkArcrole: NS.xlink,
-    xlinkHref: NS.xlink,
-    xlinkRole: NS.xlink,
-    xlinkShow: NS.xlink,
-    xlinkTitle: NS.xlink,
-    xlinkType: NS.xlink,
-    xmlBase: NS.xml,
-    xmlLang: NS.xml,
-    xmlSpace: NS.xml,
-  },
-};
-
-var CAMELIZE = /[\-\:]([a-z])/g;
-var capitalize = token => token[1].toUpperCase();
-
-ATTRS.forEach(original => {
-  var reactName = original.replace(CAMELIZE, capitalize);
-
-  SVGDOMPropertyConfig.Properties[reactName] = 0;
-  SVGDOMPropertyConfig.DOMAttributeNames[reactName] = original;
-});
+var SVGDOMPropertyConfig = {};
 
 module.exports = SVGDOMPropertyConfig;
